Sidecar injector working.
A lot of cleanup in labels - labels are expensive in k8s and create
problems when upgrading. All the 'chart' etc add little value and are
likely to create pain when we move to operator or other versions of helm
or charts. No need to tie it to helm, we may move away.

Added a simple mechanism to add hosts to the Gateway - will need to be
refined and integrated with new export/import policy.

Telemetry cleanups.